divide SpotifyApiClient into dto and client
